Refactor: refacoted components organization
