+++ package.json
@@ -20,9 +20,13 @@
   "license": "GPL-3.0",
   "dependencies": {
     "@nuxtjs/axios": "^5.3.3",
+    "@nuxtjs/dotenv": "^1.3.0",
     "@nuxtjs/proxy": "^1.2.4",
+    "config": "^2.0.1",
+    "config-webpack": "^1.0.4",
     "cookie-parser": "^1.4.3",
     "cookie-universal-nuxt": "^2.0.5",
+    "dotenv": "^6.1.0",
     "element-ui": "^2.4.8",
     "js-cookie": "^2.2.0",
     "lodash": "^4.17.11",
