+++ src/utils/event-bus.js
@@ -1,0 +1,3 @@
+import Vue from 'vue';
+const eventBus = new Vue();
+export default eventBus;
